+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'storyteller:fictionaltime',
-  version: '0.4.1',
+  version: '0.5.0',
   // Brief, one-line summary of the package.
   summary: 'Create your own linear fictional time.',
   // URL to the Git repository containing the source code for this package.
@@ -11,17 +11,18 @@ Package.describe({
 })
 
 Package.onUse(function(api) {
-  api.versionsFrom('1.9');
-  api.use('ecmascript');
+  api.versionsFrom(['1.12.1', '2.3.1']);
+  api.use(['ecmascript', 'typescript']);
   api.use('logging');
+  api.use('zodern:types@1.0.9');
 
   //export out the fictionaltime
-  api.mainModule('./fictionaltime.js');
+  api.mainModule('./fictionaltime.ts');
 })
 
 Package.onTest(function(api) {
   api.use('ecmascript');
   api.use('tinytest');
   api.use('storyteller:fictionaltime');
-  api.addFiles('fictionaltime-tests.js');
+  api.addFiles('fictionaltime-tests.ts');
 })